requirements.txt
fastapi
uvicorn
pydantic
passlib[argon2]
python-jose
psycopg2